*rage*. looks like flconsole is worthless now..
it's getting to the point where i think it's more expensive to modify it
any further and to give up on that area. scrapping the 1 sf::Text and
just going all out with glconsole (near-vanilla), porting that to use
multiple sf::Texts, one per line...that should solve all of my problems.
i hope.