sick! block breaking works on a moving plane..
now to limit it to a radius. as i said though, the algorithm to
calculate this is dumbfuck stupid. it iterates over the whole visible
screen's tilemap, instead of just the radius. since this is being done
every click..this is something i need to address ;-)

also..bit of code cleanup in this area.