a lot of moving of code going on
move msot of the GL code from Game to SpriteSheetManager, for the most
part it's all we're cared about wrt to rendering, at least right now.

later will come the tilemap rendering, but that's different. this can
be refactored later on, and that is what i am keeping in mind.

the sprite shader (sprite.{vert,frag}) will be specific, and for the
sprite shader program.

fix passing shader program to camera. perhaps i should turn camera
into a static? the the shaders could set these values themselves..
it is ugly and will not scale with the camera knowing the shader program
and passing the matrices it needs.