now using google protobuf for story binary data.
we absolutely need this to serialize the binary data for packet transfer
and also for binary (world data and other data) data storage.

basically, it means the binary data scales to any platform, and it even
has versioning..which seems like it'd be super helpful for world saves.

it's actually a very well done lib, and just involves using protoc to
compile your .pb definition files into a C++ class and header, which
then yields getter setters and even a debug string (for seeing if the data
you stored is actually good, as opposed to the fact that it's just binary
making things difficult).