attempt to send initial client data
turns out i just exposed a nasty failure in my packet serialization.
if i send a 0 as an integer anywhere in my packet, it craps out entirely.

betting it's because i call google::protobuf::Message::SerializeToString
and the ParseFromString..the examples i saw used that though to do just
this..weird.