add World::findPlayer for lookups as well as implement removeplayer
turn it into a list, much better than a vector for here. when the player
disconnects, every client will be told so, given a packet containing
playerID, they look that up, tell their own World to find that, remove
that, and then they do "something" with it (maybe), and delete it.