astyle run and initial quad tree implementation
we're going to use this for a lot of things, so i'm going to spend
a good amount of time getting it right.

presently, it'll be needed/required for torches, we'll just have a quad
tree of torches in the world and find the ones that we care about, etc.
makes for quick and easy culling, though i have to make it generic
to handle dynamic entities, which is a shockingly more complex issue
than it may seem, especially since you need to take into the rect of the
entity, and is a problem with big ones. also, outputting the contents
of the quadtree and all subnodes and entities therein is going to be
interesting...
(precommit so i can work on it with my lappy)