begin a journey of switching mouse handling over to world coordinates
the issue has existed from the beginning, because i would always just
do mouse calculations based on pixels it is on (integers), and passing
this to the server and letting it handle some calculations. Now, however,
I need to make this more elaborate as the previous approach had dead-stupid
-obvious failures.

so now the idea is to have the client get the mouse coordinates, unproject
them into the world; thus making them in world coordinates (this means
that zooming in and out just magically fucking works), then passing them
to the server. only issue being is that now i have to switch *all* of that
handling and from int to float vectors everywhere it is used. and pray
it works.

the client unprojects because the server has no notion of a camera, and
i don't believe it should.