fix threading issues with physics renderer
however, this means that when physics rendering is enabled, that the
renderer (client) thread will wait on the server thread. Since they aren't
necessarily in sync, it ends up creating a bit of lag. and it indeed means
that rendering so many objects can slow it down, and thus the server has
to wait on the client thread.

i suppose this isn't a big deal since it's just for debug, but it kinda
annoys me. going to see what i can do to help it.