doesn't compile. it did though, i swear!
it looks like qt's opengl shit is fucked. so I may have to go back to
utilizing glew/someloader.  But i'd want something better than glew because
glew is a pile of shit too.

other than that though, when it compiled it was rendering an example
gl scene beneath the qtquick ui layer. so that was cool.