cleanup some debug statements and the way i was doing the key events
honestly i really should fix this approach and truly multithreadify
the client. but i'm wondering on which scale to do this  at, and where
exactly, how to avoid lock stalls and the like.

the renderer itself would of course just be on one thread (the thread it is
on now)