try and delay renderers init..so that it is executed on the renderer thread
it was previously happening as soon as the world was created, which would
create all renderers, which would run their ctor's which wanted to do
shader compilation and buffer uploading, which obviously isn't possible.

so now on world::render we'll check if we're not init'ed.
if we're not we will init each one. The constructors for the renderers
themselves now don't do much at all in terms of GL calls (none).

though it's kind of ugly to actually be checking the init status each frame..
not like it will be noticeable in a million years.