lots of fixes, World::update() shouldn't be calling some client things
if we are not yet init'ed (renderer-wise). Also begin using qvector.

The act of me switching to qvector resulted me in instantly finding
the issue in the bt because it asserted at an ::at out of range.
STL vector did not do this. What did it do? It gave me an exception,
which I hate, and which are useless for finding the actual error without
exception handling wrapped all around your entire codebase.

asserts continue to be 1000x better, and apparently, Qt too.