implement client-side assertions for malicious player input..
though obviously it's not really malicious but accidental in this case
(as I am currently experiencing).

but comment it out for now, create a check server-side for such input
activity. I'm surprised I didn't think of it while I was writing it, but
guess I was caught up in "getting it to work".