document a lot of server functions, cleanup code rm some dead code,
pass the player pointer instead of just the player name, to the
receiveChatMessage. We'll need it later on, certainly.

begin hashing out a kickClient and kickPlayer function. former is
internal to just the server, for when we need to kick someone before
the server has created a player for said client (this step covers
very early-kick events, like invalid player name, and version mismatch).