improve version checking on server->client handshake, improve kicking
on validation failure..add the malicious intent flag to disconnect return
value. It's presumed that it'd be only from a malicious method, so other
things that get passed impossibly bad values from the client can be checked
and can kick them if need be. It' snot always malicious though, in the case
of right now, it's just a bug somewhere in the client. But most other
scenarios would be malicious, so prevent it so the server is secured properly.