create a client torch upon client world creatoin, at player position
the ideal would be to, upon player spawn, have the server put some torches
down, obviously.

But for some reason sending the item spawned command to
the client, the client tries to push_back it into the vector, and I get an odd
allocator error in std...which sounds like the allocator ran out of memory, but
that doesn't make sense to me..

So you can now see part of the tiled world, not as boring looking anymore.

Still find this odd though, as it's not like the client doesn't use
std::vectors elsewhere..