rm old experiments for particle rendering..
geometry shaders are not only extremely expensive, but only available
on modern hardware. and point sprites are hardware limited, very severely
in their number of.

Proper solution is just regular sprites, but instead of 4 verts, a good
sweet spot I've seen is 6-8 verts per particle. Reason being is that
everything is fill bound, and the vertex processor is doing nothing,
and it always needs more stuff to do..so we can use this to cull a bunch
out to reduce overdraw.

That is, when i eventually get back to working on particles..