center the escape menu properly now, so yay, that's a huge item off the
checklist of regressions/lack of feature parity.

next up is fixing the main menu to actually host/join a game..(easy),
finishing up chat dialog and connect it to make it work (med), and
create the quickbar menu with shader decorations(hard), fix mouse input
to coincide with game (med) as well as kb input. and i think that'd
be mostly the blockers for merging into master