no need for atomics here. since they're doubles anyways, the compiler
probably just compiles it to a mutex anyways..so that's where a lot
of latency is coming from I'm sure. and of course, it's undetectable
in my frametime graphs...not sure why.

i'm going to try and track down some more of this stuff, because nothing
pisses me off more than hiccups and skips in what should be a smooth
game.