actually, stop using physics debug renderer for chipmunk
it also seems to be causing gl issues on my desktop for some reason
(yes, i tested the laptop with mesa; desktop with nvidia)

but it looks like i'm going to delve into writing my own physics handling,
which has loads of options for optimizations, especially with block collision
because my current method is really dumb.

and i got thinking, that 'what does a rigid body physics engine give me'?
and the answer was basically nothing that i can think of. i guess it  was
something that *seemed* like a good idea at the time